+++ src/wifi/model/interference-helper.cc
@@ -233,7 +233,7 @@ InterferenceHelper::AppendEvent (Ptr<InterferenceHelper::Event> event)
 
 
 double
-InterferenceHelper::CalculateSnr (double signal, double noiseInterference, uint32_t channelWidth) const
+InterferenceHelper::CalculateSnr (double signal, double noiseInterference, uint8_t channelWidth) const
 {
   //thermal noise at 290K in J/s = W
   static const double BOLTZMANN = 1.3803e-23;
@@ -243,7 +243,7 @@ InterferenceHelper::CalculateSnr (double signal, double noiseInterference, uint3
   double noiseFloor = m_noiseFigure * Nt;
   double noise = noiseFloor + noiseInterference;
   double snr = signal / noise; //linear scale
-  NS_LOG_DEBUG ("bandwidth(MHz)=" << channelWidth << ", signal(W)= " << signal << ", noise(W)=" << noiseFloor << ", interference(W)=" << noiseInterference << ", snr(linear)=" << snr);
+  NS_LOG_DEBUG ("bandwidth(MHz)=" << (uint16_t)channelWidth << ", signal(W)= " << signal << ", noise(W)=" << noiseFloor << ", interference(W)=" << noiseInterference << ", snr(linear)=" << snr);
   return snr;
 }
