+++ src/wifi/model/wifi-mode.h
@@ -106,7 +106,7 @@ public:
    * If a transmission mode uses 1/2 FEC, and if its
    * data rate is 3.25Mbps, the phy rate is 6.5Mbps
    */
-  uint64_t GetPhyRate (uint32_t channelWidth, bool isShortGuardInterval, uint8_t nss) const;
+  uint64_t GetPhyRate (uint8_t channelWidth, bool isShortGuardInterval, uint8_t nss) const;
   /**
    * \param txVector the WifiTxVector of the signal
    *
@@ -124,7 +124,7 @@ public:
    *
    * \returns the data bit rate of this signal.
    */
-  uint64_t GetDataRate (uint32_t channelWidth, bool isShortGuardInterval, uint8_t nss) const;
+  uint64_t GetDataRate (uint8_t channelWidth, bool isShortGuardInterval, uint8_t nss) const;
   /**
    * \param txVector the WifiTxVector of the signal
    *
@@ -136,7 +136,7 @@ public:
    *
    * \returns the data bit rate of this non-HT or non-VHT signal.
   */
-  uint64_t GetDataRate (uint32_t channelWidth) const;
+  uint64_t GetDataRate (uint8_t channelWidth) const;
 
   /**
    * \returns the coding rate of this transmission mode
