+++ src/wifi/model/wifi-phy.h
@@ -619,7 +619,7 @@ public:
    *
    * \return true if the channel definition succeeded
    */
-  bool DefineChannelNumber (uint16_t channelNumber, WifiPhyStandard standard, uint32_t frequency, uint32_t channelWidth);
+  bool DefineChannelNumber (uint16_t channelNumber, WifiPhyStandard standard, uint32_t frequency, uint8_t channelWidth);
 
   /**
    * A pair of a ChannelNumber and WifiPhyStandard
@@ -628,7 +628,7 @@ public:
   /**
    * A pair of a center Frequency and a ChannelWidth
    */
-  typedef std::pair<uint32_t, uint32_t> FrequencyWidthPair;
+  typedef std::pair<uint32_t, uint8_t> FrequencyWidthPair;
 
   /**
    * Return the WifiChannel this WifiPhy is connected to.
@@ -1510,19 +1510,19 @@ public:
   /**
    * \return the channel width
    */
-  virtual uint32_t GetChannelWidth (void) const;
+  virtual uint8_t GetChannelWidth (void) const;
   /**
    * \param channelwidth channel width
    */
-  virtual void SetChannelWidth (uint32_t channelwidth);
+  virtual void SetChannelWidth (uint8_t channelwidth);
   /**
    * \param channelwidth channel width (in MHz) to support
    */
-  virtual void AddSupportedChannelWidth (uint32_t channelwidth);
+  virtual void AddSupportedChannelWidth (uint8_t channelwidth);
   /**
    * \return a vector containing the supported channel widths, values in MHz
    */
-  virtual std::vector<uint32_t> GetSupportedChannelWidthSet (void) const;
+  virtual std::vector<uint8_t> GetSupportedChannelWidthSet (void) const;
 
 
 protected:
@@ -1658,7 +1658,7 @@ private:
    * \param width The channel width to use
    * \return the channel number if found, zero if not
    */
-  uint16_t FindChannelNumberForFrequencyWidth (uint32_t frequency, uint32_t width) const;
+  uint16_t FindChannelNumberForFrequencyWidth (uint32_t frequency, uint8_t width) const;
   /**
    * Lookup frequency/width pair for channelNumber/standard pair
    * \param channelNumber The channel number to check
@@ -1788,7 +1788,7 @@ private:
   uint32_t m_channelCenterFrequency;   //!< Center frequency in MHz
   uint32_t m_initialFrequency;         //!< Store frequency until initialization
   bool m_frequencyChannelNumberInitialized; //!< Store initialization state
-  uint32_t m_channelWidth;             //!< Channel width
+  uint8_t m_channelWidth;             //!< Channel width
 
   double m_edThresholdW;          //!< Energy detection threshold in watts
   double   m_ccaMode1ThresholdW;  //!< Clear channel assessment (CCA) threshold in watts
@@ -1814,7 +1814,7 @@ private:
   typedef std::map<ChannelNumberStandardPair,FrequencyWidthPair> ChannelToFrequencyWidthMap;
   static ChannelToFrequencyWidthMap m_channelToFrequencyWidth;
 
-  std::vector<uint32_t> m_supportedChannelWidthSet; //!< Supported channel width
+  std::vector<uint8_t> m_supportedChannelWidthSet; //!< Supported channel width
   uint16_t              m_channelNumber;            //!< Operating channel number
   uint16_t              m_initialChannelNumber;     //!< Initial channel number
